Package refactors - VerifyingRelocationHelper*
